+++ firefly-db/src/main/java/com/firefly/db/namedparam/PreparedSqlAndValues.java
@@ -1,4 +1,4 @@
-package com.firefly.db.jdbc.helper.namedparam;
+package com.firefly.db.namedparam;
 
 import java.util.Collections;
 import java.util.List;
